Icons were added and files refactor
