+++ love/src/components/EnvironmentSummary/EnvironmentSummary.jsx
@@ -1,9 +1,12 @@
 import React, { Component } from 'react';
-import Skymap from './Skymap';
+import Skymap from './Skymap/Skymap';
 import styles from './EnvironmentSummary.module.css';
 import WindRose from 'components/icons/WindRose/WindRose';
 import SimonyiTelescope from './Cartoons/SimonyiTelescope';
 import AuxTelescope from './Cartoons/AuxTelescope';
+import BeachIcon from 'components/icons/BeachIcon/BeachIcon';
+import MountainIcon from 'components/icons/MountainIcon/MountainIcon';
+import TemperatureIcon from 'components/icons/TemperatureIcon/TemperatureIcon';
 
 export default class EnvironmentSummary extends Component {
   constructor(props) {
@@ -20,10 +23,19 @@ export default class EnvironmentSummary extends Component {
         <div className={styles.windRoseContainer}>
           <WindRose />
         </div>
+        <div className={styles.iconLeft}>
+          <BeachIcon />
+        </div>
         <div ref={this.containerRef} className={styles.telescopes}>
           <Skymap containerNode={this.containerRef?.current} width={300} height={300} className={styles.skymap} />
           <SimonyiTelescope className={styles.simonyi} />
+          <div className={styles.iconCenter}>
+            <MountainIcon />
+          </div>
           <AuxTelescope className={styles.auxTel} />
+        </div>
+        <div className={styles.iconRight}>
+          <TemperatureIcon />
         </div>
       </div>
     );
